% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_expansion.R
\name{column_expansion}
\alias{column_expansion}
\title{Column Expansion
This function performs a column expansion. A column expansion is required when
multiple pieces of data are included in a single column. This function assumes that
all seperator characters are correctly located, and the data in the column is spelled correctly.}
\usage{
column_expansion(df, column, separator)
}
\arguments{
\item{df}{dataframe}

\item{column}{name of the column for expansion}

\item{separator}{character that separates the data}
}
\value{
expanded dataframe
}
\description{
Column Expansion
This function performs a column expansion. A column expansion is required when
multiple pieces of data are included in a single column. This function assumes that
all seperator characters are correctly located, and the data in the column is spelled correctly.
}
\examples{
column_expansion(df, 1, ",")

}
