% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_row.R
\name{expand_row}
\alias{expand_row}
\title{expand row}
\usage{
expand_row(df, vector, row)
}
\arguments{
\item{df}{dataframe}

\item{vector}{vector of values to build columns for}

\item{row}{row number of dataframe}
}
\value{
updated dataframe
}
\description{
expand row
}
\examples{
expand_row(df, row_contents_vector, row_number)
}
