% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findExistingColumn.R, R/find_existing_column.R
\name{find_existing_column}
\alias{find_existing_column}
\title{findExistingColumn}
\usage{
find_existing_column(df, colname)

find_existing_column(df, colname)
}
\arguments{
\item{df}{dataframe to be checked}

\item{colName}{name of column to check dataframe for}
}
\value{
index of column or NO_COLUMN_FOUND (-999999)

index of column or NO_COLUMN_FOUND (-999999)
}
\description{
findExistingColumn

findExistingColumn
}
\examples{
findExistingColumn(insertCommas, "acetaminophen_mg")
findExistingColumn(insertCommas, "acetaminophen_mg")
}
